💅 Add #StandWithUkraine 🇺🇦 banner to org README
The https://github.com/vshymanskyy/StandWithUkraine project by a Ukrainian FOSS maintainer has gained some popularity and spearheads raising the awareness throughout the open source world. It's only fitting to show it in the FOSS effort of the Diia project.

Many projects worldwide show a variant of this banner. For example:
* https://github.com/nedbat/coveragepy#readme
* https://github.com/terraform-aws-modules
* https://github.com/platformio/platformio-core#license
* https://github.com/wasm3
* https://github.com/denysdovhan/wtfjs#readme
* https://github.com/blynkkk/blynk-library#readme
* https://github.com/cherrypy/cherrypy#readme
* and thousands more...